requirements.txt
albucore==0.0.9
albumentations==1.4.8
annotated-types==0.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
easydict==1.13
flatbuffers==24.3.25
fonttools==4.53.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.3.0
prettytable==3.10.0
protobuf==5.27.1
pycodestyle==2.11.1
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
ruff==0.4.8
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
tifffile==2024.5.22
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
